Some refactors:
- Extracted NumberFormat to JTranscLocale to reuse
- Removed JTranscStringSplit and moved split method to JTranscStrings
