+++ jtransc-rt-core/test/com/jtransc/text/JTranscStringSplitTest.java
@@ -1,5 +1,6 @@
 package com.jtransc.text;
 
+import com.jtransc.util.JTranscStrings;
 import org.junit.Test;
 
 import java.util.Arrays;
@@ -9,9 +10,9 @@ import static org.junit.Assert.assertEquals;
 public class JTranscStringSplitTest {
 	@Test
 	public void name() throws Exception {
-		assertEquals(Arrays.asList(""), Arrays.asList(JTranscStringSplit.split("", '.')));
-		assertEquals(Arrays.asList("test"), Arrays.asList(JTranscStringSplit.split("test", '.')));
-		assertEquals(Arrays.asList("hello", "world"), Arrays.asList(JTranscStringSplit.split("hello.world", '.')));
-		assertEquals(Arrays.asList("hello", "world", ""), Arrays.asList(JTranscStringSplit.split("hello.world.", '.')));
+		assertEquals(Arrays.asList(""), Arrays.asList(JTranscStrings.split("", '.')));
+		assertEquals(Arrays.asList("test"), Arrays.asList(JTranscStrings.split("test", '.')));
+		assertEquals(Arrays.asList("hello", "world"), Arrays.asList(JTranscStrings.split("hello.world", '.')));
+		assertEquals(Arrays.asList("hello", "world", ""), Arrays.asList(JTranscStrings.split("hello.world.", '.')));
 	}
 }
